+++ src/index.ts
@@ -1,3 +1,4 @@
 import '../node_modules/bulma/css/bulma.min.css'
 import '../node_modules/@mdi/font/css/materialdesignicons.min.css'
 export * from './components/atoms'
+export * from './components/molecules'